[ #31 ] Surface syntax expression/declaration
